Updated movement in progress
Started exploring new processes of mapping movement to camera where now the camera focuses on an object facing front of the player instead of the player and use another form of movement instead of the velocity as velocity proves not useful for context of moving the player forward in terms of camera facing forward. (can use rigidbody.move).