New method of Movement (Without "Front" Object)
Added new method of movement that uses the inversion of the camera's position in order to tell which direction is forwards and no longer need a second "front" object, to be used in order for the code to determine which direction is forwards facing according to the camera. Some changes were made to the camera movement code, where there is no longer an inversion in the y-axis (looking left and right) and looking in the x-axis (up and down) is more limited when looking up so that the character cannot look through the floor. Maybe at a later update, improve camera code so that it cannot move into the walls and outside of the map.