Fixed shift to sprint and ctrl to sneak
For some reason the location of the first if statement did not allow for it to be read for any reason at all, so spent some time trying to figure out the reason why but ended up just moving some of the statements around and it worked. Also camera movement when sprinting was reduced slightly in order to facilitate for camera moving out of the level.