Added obstacles for stealth section and begun camera collision by using ray cast .
Used ray cast in order to determine when there is an obstacle between the camera and the player.  The ray cast gets interrupted by the objects. when the object has a layer "wall" then there will be a relayed response. need to get camera movement done in relation to the distance between the wall and the camera so that there is no longer clipping through and between the walls.