Enemy Detection is working.
Draws a debugged area around the enemy to represent the FOV of the enemy. when the play comes into contact with this area, there is a reaction in the debug log and the debugged shapes (colours change from red to green when detecting the player). In the future, when detected, the enemy should start following the player at a certain speed but when losing sight of the player from a constant ray cast go back to patrolling.

With minor bugs with complete collision with the enemy during the patrolling where the players x value is elevated or decreased depending on if its colliding with the object directing the front of the enemy.