Some optimisations and fixes.
Tried to fix the change in camera distance when sprinting or crouching. But does not work (try to find why).
Tried to implement some sort of mechanic for crouching where the enemy finds it harder to see the player when the player is crouching.
Fixed some of the commenting as they did not accurately dictate the purpose of some aspects of the code.