Added to coin
Added rotation to coin and the ability for the coin to be collected. When it collides with another game object with the tag of player, it destroys itself and then displays a debug message saying the coin was collected. During this there is a Boolean variable that can later be used to know when the coin has been collected, and thus when it is adequate to end the game.