Added end scene for finishing the objectives.
Created an end scene for the player to reach after reaching the end of the level. This works by there being an object that, when the player passes through it, it moves them to the next scene.  However, when moving to the end scene, I had to create a new script for the mouse to now become un-stuck and visible again.  Also duplicated the main menu scene and used it for the end scene. Next need to add the game over screen for when you are caught by the security guard.