Finished adding the end for when caught.
Added the end for when caught. Changed some of the code for when the enemy is moving. So now works that the code uses the distance between the enemy and the player as a measurement and whenever the player is between the viewing distance and two other points, it will first stop and stare at the player, then begins to move towards the player as they try to escape.